--- tsconfig.json
@@ -1,6 +1,0 @@
-{
-	"extends": "@tsconfig/node14/tsconfig.json",
-	"compilerOptions": {
-		"outDir": "dist"
-	}
-}
